requirements.txt
discord.py>=2.3.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.2.1
pillow>=10.0.0
asyncio-throttle>=1.0.2
pyyaml>=6.0
aiosqlite>=0.19.0
